refactored some things and made a few constants into variables (these variables are set as constants by the caller
