--- src/main/java/apple/trees/tree/trunk/creation/Widthify.java
@@ -1,4 +1,0 @@
-package apple.trees.tree.trunk.creation;
-
-public class Widthify {
-}
