+++ src/main/java/apple/trees/tree/trunk/creation/utils/Widthify.java
@@ -1,0 +1,4 @@
+package apple.trees.tree.trunk.creation.utils;
+
+public class Widthify {
+}
